# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * BookEase AI API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
}

export type BusinessCategory =
  (typeof BusinessCategory)[keyof typeof BusinessCategory];

export const BusinessCategory = {
  clinic: "clinic",
  salon: "salon",
  gym: "gym",
  restaurant: "restaurant",
  spa: "spa",
  dental: "dental",
  barber: "barber",
  other: "other",
} as const;

export interface Business {
  id: number;
  ownerId: string;
  name: string;
  category: BusinessCategory;
  description?: string | null;
  phone?: string | null;
  email?: string | null;
  address?: string | null;
  website?: string | null;
  logo?: string | null;
  createdAt: string;
  updatedAt: string;
}

export type UpsertBusinessBodyCategory =
  (typeof UpsertBusinessBodyCategory)[keyof typeof UpsertBusinessBodyCategory];

export const UpsertBusinessBodyCategory = {
  clinic: "clinic",
  salon: "salon",
  gym: "gym",
  restaurant: "restaurant",
  spa: "spa",
  dental: "dental",
  barber: "barber",
  other: "other",
} as const;

export interface UpsertBusinessBody {
  name: string;
  category: UpsertBusinessBodyCategory;
  description?: string;
  phone?: string;
  email?: string;
  address?: string;
  website?: string;
  logo?: string;
}

export interface AvailabilitySlot {
  id: number;
  businessId: number;
  /** 0=Sunday, 1=Monday, ..., 6=Saturday */
  dayOfWeek: number;
  /** HH:MM format */
  startTime: string;
  /** HH:MM format */
  endTime: string;
  /** Duration of each appointment slot in minutes */
  slotDurationMinutes: number;
  isActive: boolean;
}

export interface CreateAvailabilitySlotBody {
  dayOfWeek: number;
  startTime: string;
  endTime: string;
  slotDurationMinutes: number;
  isActive?: boolean;
}

export type AppointmentStatus =
  (typeof AppointmentStatus)[keyof typeof AppointmentStatus];

export const AppointmentStatus = {
  pending: "pending",
  confirmed: "confirmed",
  cancelled: "cancelled",
  completed: "completed",
} as const;

export interface Appointment {
  id: number;
  businessId: number;
  customerName: string;
  customerEmail?: string | null;
  customerPhone?: string | null;
  service?: string | null;
  appointmentDate: string;
  notes?: string | null;
  status: AppointmentStatus;
  createdAt: string;
}

export interface CreateAppointmentBody {
  customerName: string;
  customerEmail?: string;
  customerPhone?: string;
  service?: string;
  appointmentDate: string;
  notes?: string;
}

export type UpdateAppointmentStatusBodyStatus =
  (typeof UpdateAppointmentStatusBodyStatus)[keyof typeof UpdateAppointmentStatusBodyStatus];

export const UpdateAppointmentStatusBodyStatus = {
  pending: "pending",
  confirmed: "confirmed",
  cancelled: "cancelled",
  completed: "completed",
} as const;

export interface UpdateAppointmentStatusBody {
  status: UpdateAppointmentStatusBodyStatus;
}

export interface DashboardSummary {
  totalAppointments: number;
  pendingAppointments: number;
  confirmedAppointments: number;
  completedAppointments: number;
  cancelledAppointments: number;
  todayAppointments: number;
  thisWeekAppointments: number;
}

export interface SendChatMessageBody {
  message: string;
  /** Client session ID for conversation continuity */
  sessionId: string;
}

export interface OpenaiConversation {
  id: number;
  title: string;
  createdAt: string;
}

export interface OpenaiMessage {
  id: number;
  conversationId: number;
  role: string;
  content: string;
  createdAt: string;
}

export interface CreateOpenaiConversationBody {
  title: string;
}

export interface SendOpenaiMessageBody {
  content: string;
}

export interface OpenaiConversationWithMessages {
  id: number;
  title: string;
  createdAt: string;
  messages: OpenaiMessage[];
}

export interface OpenaiError {
  error: string;
}

export type ListMyAppointmentsParams = {
  status?: ListMyAppointmentsStatus;
  /**
   * Filter by date (YYYY-MM-DD)
   */
  date?: string;
};

export type ListMyAppointmentsStatus =
  (typeof ListMyAppointmentsStatus)[keyof typeof ListMyAppointmentsStatus];

export const ListMyAppointmentsStatus = {
  pending: "pending",
  confirmed: "confirmed",
  cancelled: "cancelled",
  completed: "completed",
} as const;