# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: BookEase AI API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /businesses/me:
    get:
      operationId: getMyBusiness
      tags: [businesses]
      summary: Get the current user's business profile
      responses:
        "200":
          description: Business profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Business"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    put:
      operationId: upsertMyBusiness
      tags: [businesses]
      summary: Create or update the current user's business profile
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpsertBusinessBody"
      responses:
        "200":
          description: Updated business
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Business"
  /businesses/{businessId}:
    get:
      operationId: getBusinessById
      tags: [businesses]
      summary: Get a business by ID (public, for customer chatbot)
      parameters:
        - name: businessId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Business profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Business"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /businesses/me/availability:
    get:
      operationId: listMyAvailability
      tags: [availability]
      summary: List availability slots for current user's business
      responses:
        "200":
          description: List of availability slots
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/AvailabilitySlot"
    post:
      operationId: createAvailabilitySlot
      tags: [availability]
      summary: Create a new availability slot
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateAvailabilitySlotBody"
      responses:
        "201":
          description: Created slot
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AvailabilitySlot"
  /businesses/me/availability/{slotId}:
    put:
      operationId: updateAvailabilitySlot
      tags: [availability]
      summary: Update an availability slot
      parameters:
        - name: slotId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateAvailabilitySlotBody"
      responses:
        "200":
          description: Updated slot
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AvailabilitySlot"
    delete:
      operationId: deleteAvailabilitySlot
      tags: [availability]
      summary: Delete an availability slot
      parameters:
        - name: slotId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
  /businesses/{businessId}/availability:
    get:
      operationId: listBusinessAvailability
      tags: [availability]
      summary: List availability for a business (public, for customer chatbot)
      parameters:
        - name: businessId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: List of availability slots
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/AvailabilitySlot"
  /businesses/me/appointments:
    get:
      operationId: listMyAppointments
      tags: [appointments]
      summary: List all appointments for current user's business
      parameters:
        - name: status
          in: query
          schema:
            type: string
            enum: [pending, confirmed, cancelled, completed]
        - name: date
          in: query
          schema:
            type: string
            description: Filter by date (YYYY-MM-DD)
      responses:
        "200":
          description: List of appointments
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Appointment"
  /businesses/me/appointments/{appointmentId}:
    get:
      operationId: getMyAppointment
      tags: [appointments]
      summary: Get a specific appointment
      parameters:
        - name: appointmentId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Appointment
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Appointment"
    patch:
      operationId: updateAppointmentStatus
      tags: [appointments]
      summary: Update appointment status
      parameters:
        - name: appointmentId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateAppointmentStatusBody"
      responses:
        "200":
          description: Updated appointment
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Appointment"
  /businesses/{businessId}/appointments:
    post:
      operationId: createAppointment
      tags: [appointments]
      summary: Book an appointment (public, via chatbot)
      parameters:
        - name: businessId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateAppointmentBody"
      responses:
        "201":
          description: Created appointment
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Appointment"
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error
    Business:
      type: object
      properties:
        id:
          type: integer
        ownerId:
          type: string
        name:
          type: string
        category:
          type: string
          enum: [clinic, salon, gym, restaurant, spa, dental, barber, other]
        description:
          type: string
          nullable: true
        phone:
          type: string
          nullable: true
        email:
          type: string
          nullable: true
        address:
          type: string
          nullable: true
        website:
          type: string
          nullable: true
        logo:
          type: string
          nullable: true
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required:
        - id
        - ownerId
        - name
        - category
        - createdAt
        - updatedAt
    UpsertBusinessBody:
      type: object
      properties:
        name:
          type: string
        category:
          type: string
          enum: [clinic, salon, gym, restaurant, spa, dental, barber, other]
        description:
          type: string
        phone:
          type: string
        email:
          type: string
        address:
          type: string
        website:
          type: string
        logo:
          type: string
      required:
        - name
        - category
    AvailabilitySlot:
      type: object
      properties:
        id:
          type: integer
        businessId:
          type: integer
        dayOfWeek:
          type: integer
          description: "0=Sunday, 1=Monday, ..., 6=Saturday"
        startTime:
          type: string
          description: "HH:MM format"
        endTime:
          type: string
          description: "HH:MM format"
        slotDurationMinutes:
          type: integer
          description: Duration of each appointment slot in minutes
        isActive:
          type: boolean
      required:
        - id
        - businessId
        - dayOfWeek
        - startTime
        - endTime
        - slotDurationMinutes
        - isActive
    CreateAvailabilitySlotBody:
      type: object
      properties:
        dayOfWeek:
          type: integer
        startTime:
          type: string
        endTime:
          type: string
        slotDurationMinutes:
          type: integer
        isActive:
          type: boolean
      required:
        - dayOfWeek
        - startTime
        - endTime
        - slotDurationMinutes
    Appointment:
      type: object
      properties:
        id:
          type: integer
        businessId:
          type: integer
        customerName:
          type: string
        customerEmail:
          type: string
          nullable: true
        customerPhone:
          type: string
          nullable: true
        service:
          type: string
          nullable: true
        appointmentDate:
          type: string
          format: date-time
        notes:
          type: string
          nullable: true
        status:
          type: string
          enum: [pending, confirmed, cancelled, completed]
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - businessId
        - customerName
        - appointmentDate
        - status
        - createdAt
    CreateAppointmentBody:
      type: object
      properties:
        customerName:
          type: string
        customerEmail:
          type: string
        customerPhone:
          type: string
        service:
          type: string
        appointmentDate:
          type: string
          format: date-time
        notes:
          type: string
      required:
        - customerName
        - appointmentDate
    UpdateAppointmentStatusBody:
      type: object
      properties:
        status:
          type: string
          enum: [pending, confirmed, cancelled, completed]
      required:
        - status
...