# lib/api-zod/src/generated/types/availabilitySlot.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * BookEase AI API specification
 * OpenAPI spec version: 0.1.0
 */

export interface AvailabilitySlot {
  id: number;
  businessId: number;
  /** 0=Sunday, 1=Monday, ..., 6=Saturday */
  dayOfWeek: number;
  /** HH:MM format */
  startTime: string;
  /** HH:MM format */
  endTime: string;
  /** Duration of each appointment slot in minutes */
  slotDurationMinutes: number;
  isActive: boolean;
}