# lib/api-zod/src/generated/types/business.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * BookEase AI API specification
 * OpenAPI spec version: 0.1.0
 */
import type { BusinessCategory } from "./businessCategory";

export interface Business {
  id: number;
  ownerId: string;
  name: string;
  category: BusinessCategory;
  description?: string | null;
  phone?: string | null;
  email?: string | null;
  address?: string | null;
  website?: string | null;
  logo?: string | null;
  createdAt: Date;
  updatedAt: Date;
}