# lib/api-zod/src/generated/types/createAppointmentBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * BookEase AI API specification
 * OpenAPI spec version: 0.1.0
 */

export interface CreateAppointmentBody {
  customerName: string;
  customerEmail?: string;
  customerPhone?: string;
  service?: string;
  appointmentDate: Date;
  notes?: string;
}