# lib/api-zod/src/generated/types/updateAppointmentStatusBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * BookEase AI API specification
 * OpenAPI spec version: 0.1.0
 */
import type { UpdateAppointmentStatusBodyStatus } from "./updateAppointmentStatusBodyStatus";

export interface UpdateAppointmentStatusBody {
  status: UpdateAppointmentStatusBodyStatus;
}