# lib/api-zod/src/generated/types/upsertBusinessBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * BookEase AI API specification
 * OpenAPI spec version: 0.1.0
 */
import type { UpsertBusinessBodyCategory } from "./upsertBusinessBodyCategory";

export interface UpsertBusinessBody {
  name: string;
  category: UpsertBusinessBodyCategory;
  description?: string;
  phone?: string;
  email?: string;
  address?: string;
  website?: string;
  logo?: string;
}